+++ Cargo.toml
@@ -4,4 +4,5 @@ members = [
   "dcl-common",
   "catalyst",
   "scene-compiler",
+  "ecs-v1"
 ]
