+++ ecs-v1/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "dcl2d-ecs-v1"
+version = "0.0.0"
+edition = "2021"
+resolver = "2"
+
+[dependencies]
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
+rmp-serde = "1.1.1"
+dcl-common = { path = "../dcl-common" }
