+++ ecs-v1/fixtures/components/alpha_collider.json
@@ -1,0 +1,5 @@
+{
+  "sprite": "a_pixel.png",
+  "channel": "R",
+  "anchor": "BottomCenter"
+}
