+++ ecs-v1/fixtures/components/alpha_collider.mp
@@ -1,0 +1,1 @@
+“«a_pixel.png¡R¬BottomCenter
