+++ ecs-v1/src/anchor.rs
@@ -1,0 +1,15 @@
+use serde::{Serialize, Deserialize};
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub enum Anchor{
+    Center,
+    BottomLeft,
+    BottomCenter,
+    BottomRight,
+    CenterLeft,
+    CenterRight,
+    TopLeft,
+    TopCenter,
+    TopRight,
+    Custom([i16; 2]),
+}
