+++ ecs-v1/src/color.rs
@@ -1,0 +1,9 @@
+use serde::{Serialize, Deserialize};
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub enum Channel {
+  R,
+  G,
+  B,
+  A
+}
