+++ ecs-v1/src/component.rs
@@ -1,0 +1,11 @@
+use serde::{Serialize, Deserialize};
+use crate::components::*;
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub enum Component{
+    Transform(Transform),
+    SpriteRenderer(SpriteRenderer),
+    CircleCollider(CircleCollider),
+    BoxCollider(BoxCollider),
+    AlphaCollider(AlphaCollider),
+}
