+++ ecs-v1/src/components/circle_collider.rs
@@ -1,0 +1,7 @@
+use serde::{Serialize, Deserialize};
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct CircleCollider {
+    pub center: [i16; 2],
+    pub radius: i32,
+}
