+++ ecs-v1/src/entity.rs
@@ -1,0 +1,10 @@
+use serde::{Serialize, Deserialize};
+use crate::Component;
+
+#[derive(Deserialize, Serialize, Debug, Clone)]
+pub struct Entity {
+    #[serde(skip)]
+    pub id: usize,
+    pub name: String,
+    pub components: Vec<Component>,
+}
