+++ ecs-v1/src/lib.rs
@@ -1,0 +1,13 @@
+mod anchor;
+mod component;
+mod entity;
+mod scene;
+
+pub mod color;
+pub mod components;
+pub mod test_utils;
+
+pub use anchor::Anchor;
+pub use component::Component;
+pub use entity::Entity;
+pub use scene::Scene;
