+++ scene-compiler/Cargo.toml
@@ -15,6 +15,7 @@ path = "src/lib.rs"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 dcl-common = { path = "../dcl-common" }
+dcl2d-ecs-v1 = { path = "../ecs-v1" }
 rmp-serde = "1.1.1"
 tempdir = "0.3"
 clap = { version = "4.0.17", features = ["derive"] }
